Change link at bottom of page
The link as it currently exists will take the user to 'https://docs.lando.dev/config%2Ftooling.html' when just using the left mouse button, this leads to a 404 not found page. When opening as a new tab or window it will take the user to the correct page: https://docs.lando.dev/core/v3/tooling.html.

This update simply changes the link to always send the user to the correct page for more tooling information.